requirements.txt
flask
python-dotenv
twilio
GPT4All